requirements.txt
pyspark
scikit-learn
psycopg2
kafka-python
grafana-api
